SkylineTester: Fix non-persisted language combo boxes in Forms and Tutorials.
Former-commit-id: bf7e8a464e03c7d924bf62b38a59ed5dd8a43d80

+++ pwiz_tools/Skyline/SkylineTester/Main.cs
@@ -150,7 +150,8 @@ namespace SkylineTester
         {
             //MemoryChartWindow.Start("TestRunnerMemory.log");
             TestsRun = 0;
-            Try.Multi<Exception>(() => Directory.Delete(_resultsDir, true), 4, false);
+            if (Directory.Exists(_resultsDir))
+                Try.Multi<Exception>(() => Directory.Delete(_resultsDir, true), 4, false);
 
             var testRunner = Path.Combine(GetSelectedBuildDir(), "TestRunner.exe");
             _testRunnerIndex = commandShell.Add(
